Fix type conflict

+++ src/Model/FileParser.php
@@ -151,7 +151,7 @@ class FileParser
         $column->setFixed(true);
         if ($column->getType() instanceof StringType) {
             $bracketExpressionItem = $columnTypeItem->subTree->getItem('bracket_expression');
-            $column->setLength($bracketExpressionItem ? trim($bracketExpressionItem->getBaseExpr(), '()') : null);
+            $column->setLength($bracketExpressionItem ? (int)trim($bracketExpressionItem->getBaseExpr(), '()') : null);
             if ($fieldType === 'geometry') {
                 $column->setFixed(false);
             } else {
